agent/sqlib: make the sqlib private by moving it into internal
