+++ agent/internal/agent.go
@@ -24,10 +24,10 @@ import (
 	"github.com/sqreen/go-agent/agent/internal/plog"
 	"github.com/sqreen/go-agent/agent/internal/record"
 	"github.com/sqreen/go-agent/agent/internal/rule"
-	"github.com/sqreen/go-agent/agent/sqlib/sqerrors"
-	"github.com/sqreen/go-agent/agent/sqlib/sqsafe"
-	"github.com/sqreen/go-agent/agent/sqlib/sqsanitize"
-	"github.com/sqreen/go-agent/agent/sqlib/sqtime"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqerrors"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqsafe"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqsanitize"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqtime"
 	"github.com/sqreen/go-agent/agent/types"
 	"github.com/sqreen/go-agent/sdk"
 	"golang.org/x/xerrors"
