+++ agent/internal/metrics.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/sqreen/go-agent/agent/internal/metrics"
 	"github.com/sqreen/go-agent/agent/internal/record"
-	"github.com/sqreen/go-agent/agent/sqlib/sqerrors"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqerrors"
 )
 
 func (a *Agent) AddUserEvent(event record.UserEventFace) {
