+++ agent/internal/rule/callback/waf.go
@@ -17,8 +17,8 @@ import (
 	"github.com/sqreen/go-agent/agent/internal/binding-accessor"
 	"github.com/sqreen/go-agent/agent/internal/httphandler"
 	"github.com/sqreen/go-agent/agent/internal/record"
-	"github.com/sqreen/go-agent/agent/sqlib/sqerrors"
-	"github.com/sqreen/go-agent/agent/sqlib/sqhook"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqerrors"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqhook"
 	"github.com/sqreen/go-libsqreen/waf"
 	waf_types "github.com/sqreen/go-libsqreen/waf/types"
 )
