+++ agent/internal/rule/callback/write-custom-error-page.go
@@ -10,8 +10,8 @@ import (
 	"net/http"
 
 	"github.com/sqreen/go-agent/agent/internal/backend/api"
-	"github.com/sqreen/go-agent/agent/sqlib/sqerrors"
-	"github.com/sqreen/go-agent/agent/sqlib/sqhook"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqerrors"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqhook"
 )
 
 // NewWriteCustomErrorPageCallbacks returns the native prolog and epilog
