+++ agent/internal/rule/callback/write-http-redirection.go
@@ -11,8 +11,8 @@ import (
 	"net/url"
 
 	"github.com/sqreen/go-agent/agent/internal/backend/api"
-	"github.com/sqreen/go-agent/agent/sqlib/sqerrors"
-	"github.com/sqreen/go-agent/agent/sqlib/sqhook"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqerrors"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqhook"
 )
 
 // NewWriteHTTPRedirectionCallbacks returns the native prolog and epilog
