+++ agent/internal/sqlib/sqhook/hook.go
@@ -56,9 +56,9 @@ import (
 	"sync/atomic"
 	"unsafe"
 
-	"github.com/sqreen/go-agent/agent/sqlib/sqerrors"
-	"github.com/sqreen/go-agent/agent/sqlib/sqhook/internal"
-	"github.com/sqreen/go-agent/agent/sqlib/sqsafe"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqerrors"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqhook/internal"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqsafe"
 )
 
 //go:linkname _sqreen_atomic_load_pointer _sqreen_atomic_load_pointer
