+++ agent/internal/sqlib/sqhook/test/testdata/hello-example/main.go
@@ -7,7 +7,7 @@ package main
 import (
 	"fmt"
 
-	"github.com/sqreen/go-agent/agent/sqlib/sqhook/test/testdata/helpers"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqhook/test/testdata/helpers"
 )
 
 // no params nor results
