+++ agent/internal/sqlib/sqhook/test/testdata/hello-example/test.go
@@ -5,7 +5,7 @@
 package main
 
 import (
-	"github.com/sqreen/go-agent/agent/sqlib/sqhook/test/testdata/helpers"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqhook/test/testdata/helpers"
 )
 
 func init() {
