+++ agent/internal/sqlib/sqhook/test/testdata/hello-world/test.go
@@ -7,7 +7,7 @@
 package main
 
 import (
-	"github.com/sqreen/go-agent/agent/sqlib/sqhook/test/testdata/helpers"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqhook/test/testdata/helpers"
 )
 
 func init() {
