+++ agent/internal/sqlib/sqsafe/call_test.go
@@ -7,35 +7,32 @@ package sqsafe_test
 import (
 	"testing"
 
-	"github.com/sqreen/go-agent/agent/sqlib/sqsafe"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqsafe"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/xerrors"
 )
 
-func TestGo(t *testing.T) {
+func TestCall(t *testing.T) {
 	t.Run("without error", func(t *testing.T) {
-		ch := sqsafe.Go(func() error {
+		err := sqsafe.Call(func() error {
 			return nil
 		})
-		err := <-ch
 		require.NoError(t, err)
 	})
 
 	t.Run("with a regular error", func(t *testing.T) {
-		ch := sqsafe.Go(func() error {
+		err := sqsafe.Call(func() error {
 			return xerrors.New("oops")
 		})
-		err := <-ch
 		require.Error(t, err)
 		require.Equal(t, "oops", err.Error())
 	})
 
 	t.Run("with a panic string error", func(t *testing.T) {
-		ch := sqsafe.Go(func() error {
+		err := sqsafe.Call(func() error {
 			panic("oops")
 			return nil
 		})
-		err := <-ch
 		require.Error(t, err)
 		var panicErr *sqsafe.PanicError
 		require.Error(t, err)
@@ -44,24 +41,23 @@ func TestGo(t *testing.T) {
 	})
 
 	t.Run("with a panic error", func(t *testing.T) {
-		ch := sqsafe.Go(func() error {
-			panic(xerrors.New("oops"))
+		origErr := xerrors.New("oops")
+		err := sqsafe.Call(func() error {
+			panic(origErr)
 			return nil
 		})
-		err := <-ch
 		require.Error(t, err)
 		var panicErr *sqsafe.PanicError
 		require.Error(t, err)
 		require.True(t, xerrors.As(err, &panicErr))
-		require.Equal(t, "oops", panicErr.Err.Error())
+		require.True(t, xerrors.Is(err, origErr))
 	})
 
 	t.Run("with another panic argument type", func(t *testing.T) {
-		ch := sqsafe.Go(func() error {
+		err := sqsafe.Call(func() error {
 			panic(33.7)
 			return nil
 		})
-		err := <-ch
 		require.Error(t, err)
 		var panicErr *sqsafe.PanicError
 		require.Error(t, err)
@@ -70,12 +66,11 @@ func TestGo(t *testing.T) {
 	})
 
 	t.Run("with a nil panic argument value", func(t *testing.T) {
-		ch := sqsafe.Go(func() error {
+		err := sqsafe.Call(func() error {
 			// This case cannot be differentiated yet.
 			panic(nil)
 			return xerrors.New("oops")
 		})
-		err := <-ch
 		require.NoError(t, err)
 	})
 }
