+++ agent/internal/sqlib/sqsafe/go_test.go
@@ -7,32 +7,35 @@ package sqsafe_test
 import (
 	"testing"
 
-	"github.com/sqreen/go-agent/agent/sqlib/sqsafe"
+	"github.com/sqreen/go-agent/agent/internal/sqlib/sqsafe"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/xerrors"
 )
 
-func TestCall(t *testing.T) {
+func TestGo(t *testing.T) {
 	t.Run("without error", func(t *testing.T) {
-		err := sqsafe.Call(func() error {
+		ch := sqsafe.Go(func() error {
 			return nil
 		})
+		err := <-ch
 		require.NoError(t, err)
 	})
 
 	t.Run("with a regular error", func(t *testing.T) {
-		err := sqsafe.Call(func() error {
+		ch := sqsafe.Go(func() error {
 			return xerrors.New("oops")
 		})
+		err := <-ch
 		require.Error(t, err)
 		require.Equal(t, "oops", err.Error())
 	})
 
 	t.Run("with a panic string error", func(t *testing.T) {
-		err := sqsafe.Call(func() error {
+		ch := sqsafe.Go(func() error {
 			panic("oops")
 			return nil
 		})
+		err := <-ch
 		require.Error(t, err)
 		var panicErr *sqsafe.PanicError
 		require.Error(t, err)
@@ -41,23 +44,24 @@ func TestCall(t *testing.T) {
 	})
 
 	t.Run("with a panic error", func(t *testing.T) {
-		origErr := xerrors.New("oops")
-		err := sqsafe.Call(func() error {
-			panic(origErr)
+		ch := sqsafe.Go(func() error {
+			panic(xerrors.New("oops"))
 			return nil
 		})
+		err := <-ch
 		require.Error(t, err)
 		var panicErr *sqsafe.PanicError
 		require.Error(t, err)
 		require.True(t, xerrors.As(err, &panicErr))
-		require.True(t, xerrors.Is(err, origErr))
+		require.Equal(t, "oops", panicErr.Err.Error())
 	})
 
 	t.Run("with another panic argument type", func(t *testing.T) {
-		err := sqsafe.Call(func() error {
+		ch := sqsafe.Go(func() error {
 			panic(33.7)
 			return nil
 		})
+		err := <-ch
 		require.Error(t, err)
 		var panicErr *sqsafe.PanicError
 		require.Error(t, err)
@@ -66,11 +70,12 @@ func TestCall(t *testing.T) {
 	})
 
 	t.Run("with a nil panic argument value", func(t *testing.T) {
-		err := sqsafe.Call(func() error {
+		ch := sqsafe.Go(func() error {
 			// This case cannot be differentiated yet.
 			panic(nil)
 			return xerrors.New("oops")
 		})
+		err := <-ch
 		require.NoError(t, err)
 	})
 }
